% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_chapter.R
\name{run_chapter}
\alias{run_chapter}
\title{Runs chapter}
\usage{
run_chapter(chapter_path)
}
\arguments{
\item{chapter_path}{a book chapter like 2.2 or 3}
}
\value{

}
\description{
Runs chapter
}
\examples{

\dontrun{
   run_chapter(3)
   run_chapter(2.2)
}
}
